Problem with 'linebreak'
set linebreak does not break the line, if the line contains a tab
and the line size is only as long as the screen (counting tab as 1).
(although because of the tab size, the screen line will be longer
than the screen width, thus is should be broken).
Example:
vim -u NONE -N -c "put =repeat('a', winwidth(0))" -c
's/^.\{8}\zs./\t/' -c ':set linebreak'

https://groups.google.com/d/msg/vim_dev/siAcVia-5YY/ngKUMB-galYJ