Add :domodelines command
This patch does the following: Skip reading modelines for
`:doautocmd User` if the autocommand is not actually executed.
While this is a nice performance improvement, this might be backwards
incompatible, if a one used `:doautocmd foobar` to rely on the fact,
that this will re-evaluate the modelines. Therefore, the new command:
`:domodelines` is also introduced with this patch.

See the discussion: https://groups.google.com/d/msg/vim_dev/DidKMDAsppw/vUC9CF9VNFsJ