open folds when search and incsearch is set
When searching and incsearch is set, the cursor will only stay on the
closed folds and not open the folds. This means, one cannot see context.

So this patch makes it so, that even in incsearch mode, the folds will
be opened and when moving out of the folds after typing the next char,
close the folds again.

(the actual search after typing enter works as expected).

Requested at:
https://groups.google.com/d/msg/vim_use/f-GF4glSU4g/DLCo9yq54SIJ