Fix issue 212
This patch fixes the following problem:

Consider this buffer content:
------------
a b c
a b c
------------
" map . to this:
nnoremap . :call feedkeys(".", "n")<cr>
" At line one do this:
qqdw.ifoo<esc>q
" register q should contain:
dw.ifoo
" replay macro at line two:
@q
" line 2 should look like line 1:
fooc
" and not like this.
fofooob c

This happens, because when the call to feedkeys() is executed, the
'.' will be appended to the last typed/stuffed chars while in fact
it should be placed at the current position, since you want to repeat
previous command and not the command that happens to be the last one
typed (in the mapping).

The fix is to put the '.' at the current typebuf.offset
rather at the end. I am not sure, if this should happen only when
'.' is the only char or whenever the mapped key string just starts
with '.' (but I tend to believe the second case).