Make Ctrl-N/P jump to next/previous search match
Currently, you cannot move from one match to the next match
when doing a search '/' or '?'.
This patch adds the functionality to use 'Ctrl-N' to move the
cursor to the next match, if 'insearch' is set. Similarily 'Ctrl-P' will
move to the previous match.

Also c_CTRL-N and c_CTRL-P are already used to move within in history of
search patterns, I have for now made them something different in search
mode, when incsearch is set. This is because c_CTRL-L already does
something useful in search mode and second, because Ctrl-N and
Ctrl-P are already used to select next/previous match in completion mode
so it seems logically to also extend their use in search mode.

Pressing Enter will leave you at the match, while pressing ESC will
return you to the position from where you started.

https://groups.google.com/d/msg/vim_dev/yEm6YZBfOVQ/UBeUXrdLAQAJ