updated a couple of patches
most of those patches have been lying around for too long. Not sure if
they still apply (as probably most of those patches in this repository
won't apply cleanly).

If you are using one of the patches here and notice it doesn't apply or
there is a bug, please let me know.